\name{plot_data}
\alias{plot_data}
\docType{data}
\title{
plot_data
}
\description{
LD from HapMap CEU and JPT populations
}

\format{
  The format is:
 num [1:103, 1:103] NA 0.007 0.015 0.449 0.007 1 0.454 0.002 0.425 0.786 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:103] "17766858" "17773336" "17773458" "17778418" ...
  ..$ : chr [1:103] "17766858" "17773336" "17773458" "17778418" ...
}
\details{
103 SNPS
}
\source{
HapMap
}
\references{HapMap
}
\examples{
data(plot_data)
## maybe str(plot_data) ; plot(plot_data) ...
}
\keyword{datasets}
